Lib Coverage Increase and Naming Consistency
